pyproject.toml
[tool.poetry]
name = "djangotestproject"
version = "0.1.0"
description = ""
authors = ["UTVova <utvova1@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9"
djangorestframework = "^3.12.1"
Django = "^2"
django-stdimage = "^5.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.8.2"
flake8 = "^3.8.4"
black = "^20.8b1"
seed-isort-config = "^2.2.0"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
use_parentheses = true
multi_line_output = 3
force_grid_wrap = 0
include_trailing_comma = true